Move onboarding redirect to app home redirect
