+++ src/App.spec.jsx
@@ -85,6 +85,7 @@ describe('App', () => {
   function setup(
     { termsOfServicePage = false } = { termsOfServicePage: false }
   ) {
+    const mockMutationVariables = jest.fn()
     useFlags.mockReturnValue({
       termsOfServicePage,
     })
@@ -104,8 +105,26 @@ describe('App', () => {
             },
           })
         )
-      )
+      ),
+      graphql.mutation('updateDefaultOrganization', (req, res, ctx) => {
+        mockMutationVariables(req.variables)
+
+        return res(
+          ctx.status(200),
+          ctx.data({
+            updateDefaultOrganization: {
+              defaultOrg: {
+                username: 'criticalRole',
+              },
+            },
+          })
+        )
+      })
     )
+
+    return {
+      mockMutationVariables,
+    }
   }
 
   const cloudFullRouterCases = [
@@ -266,6 +285,26 @@ describe('App', () => {
         expected: {
           page: /LoginPage/i,
           location: '/login/gh',
+        },
+      },
+    ],
+    [
+      {
+        testLabel: 'provider page',
+        pathname: '/gh',
+        expected: {
+          page: /OwnerPage/i,
+          location: '/gh/CodecovUser',
+        },
+      },
+    ],
+    [
+      {
+        testLabel: 'provider page with request params',
+        pathname: '/gh?setup_action=request',
+        expected: {
+          page: /OwnerPage/i,
+          location: '/gh/CodecovUser',
         },
       },
     ],
@@ -473,4 +512,39 @@ describe('App', () => {
       })
     }
   )
+
+  describe('feature flag is on and set up action param is request', () => {
+    it('renders children', async () => {
+      setup()
+
+      render(<App />, {
+        wrapper: wrapper(['/gh/CodecovUser?setup_action=request']),
+      })
+
+      await waitFor(() => expect(testLocation.pathname).toBe('/gh/CodecovUser'))
+
+      await waitFor(() =>
+        expect(testLocation.search).toEqual('?setup_action=request')
+      )
+    })
+  })
+
+  it('fires update default org mutation', async () => {
+    const { mockMutationVariables } = setup()
+
+    render(<App />, {
+      wrapper: wrapper(['/gh?setup_action=request']),
+    })
+
+    await waitFor(() => queryClient.isFetching)
+    await waitFor(() => !queryClient.isFetching)
+
+    await waitFor(() =>
+      expect(mockMutationVariables).toHaveBeenCalledWith({
+        input: {
+          username: 'CodecovUser',
+        },
+      })
+    )
+  })
 })
